Optimize count_all_messages()
When counting the exact number of messages, `dm.Message.grep()[0]` is used to
get the total. Since no arguments are given, no filters are added to the
Message.query. However, with the default defer=True argument, the grep method
returns the total count, the number of pages, and `Message.query.all()`.
Thankfully, we can pass defer=True and it won't call all() on the query.