Fix two bugs.
First, the value of `start` could be None here, which causes the error seen in #206.
Furthermore, it turns out that the `flask` module has no `BadRequest` exception
there.  It is really in `werkzeug.exceptions`.  Use that.

Lastly, this adds a test, which we really should've added in #197 and *really*
should've added in #204.