add transaction handling
Right now, datagrepper uses one long transaction for the life of the
process. This makes database management more difficult (it holds locks),
and it may be causing a memory leak (since sqlalchemy objects are never
released). This change causes the transaction to be rolled back at the
end of every request, and the session to be removed. A new session will
be created and transaction started at the beginning of the next
request.