Fix timezone ambiguity in tests
I found when trying to validate a PR locally that the tests were not
passing on the develop branch.  This should remove timezone
ambiguity by setting UTC everywhere in the tests.

I found also that `time.mktime(datetime.timetuple())` was an old
idiom for converting a datetime to a timestamp that had a new
replacement in `datetime.timestamp(datetime)`.