Make the approximate message count work under timescaledb
The first thing this commit does is move the setting of the default
value of the DATAGREPPER_APPROXIMATE_COUNT config setting to the
default_config file rather than hard coding it in app.py

Secondly, this fixes the approximate count not working on timescaledb.
Previously we were calculating the approx number of messages with

SELECT reltuples FROM pg_class WHERE relname = 'messages';

as a workaround for slow actual counting of these rows when we have
millions of messages. However, this query does not work with
timescaledb, as reported in:

https://github.com/timescale/timescaledb/issues/525

and implemented in as the hypertable_approximate_row_count convenience
function in:

https://github.com/timescale/timescaledb/pull/539

ypertable_approximate_row_count was renamed to approximate_row_count()
in timescaledb 2.0, which is what is used in this commit.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>